+++ LpgConsumptionCostCalculator.Web/Views/Account/Claims.cshtml
@@ -1,0 +1,26 @@
+﻿@model IEnumerable<System.Security.Claims.Claim>
+
+@{
+    ViewBag.Title = "View claims";
+}
+
+<h2>@ViewBag.Title</h2>
+
+<dl class="dl-horizontal">
+    @foreach (var claim in Model)
+    {
+        <dt title="@claim.Type">
+            @claim.Type
+            <button type="button"
+                    class="btn btn-link btn-xs"
+                    aria-label="Copy to clipboard"
+                    title="Copy to clipboard"
+                    data-clipboard-text="@claim.Value">
+                <span class="glyphicon glyphicon glyphicon-copy" aria-hidden="true"></span>
+            </button>
+        </dt>
+
+        <dd>@claim.Value</dd>
+    }
+</dl>
+
