+++ LpgConsumptionCostCalculator.Web/Views/Account/Login.cshtml
@@ -1,0 +1,30 @@
+﻿@using System.Configuration;
+@{ 
+    ViewData["Title"] = "Login"; 
+}
+
+<script src="https://global.oktacdn.com/okta-signin-widget/3.2.1/js/okta-sign-in.min.js" type="text/javascript"></script>
+<link href="https://global.oktacdn.com/okta-signin-widget/3.2.1/css/okta-sign-in.min.css" type="text/css" rel="stylesheet" />
+
+<div id="widget"></div>
+
+<form method="POST" asp-action="Login">
+    <input type="hidden" name="sessionToken" id="hiddenSessionTokenField" />
+    @Html.AntiForgeryToken()
+</form>
+
+<script type="text/javascript">
+    const signIn = new OktaSignIn({
+        baseUrl: '@ConfigurationManager.AppSettings["okta:OktaDomain"]'
+    });
+
+    signIn.renderEl({ el: '#widget' }, (res) => {
+        var sessionTokenField = $("#hiddenSessionTokenField");
+        sessionTokenField.val(res.session.token);
+        var form = sessionTokenField.parent();
+        form.submit();
+    }, (err) => {
+        console.error(err);
+    });
+</script>
+
