+++ LpgConsumptionCostCalculator.Web/Web.config
@@ -13,9 +13,11 @@
     <add key="okta:ClientId" value="0oa40y95m6pB1WAEN4x6" />
     <add key="okta:ClientSecret" value="KY6UYNYoGTPL_W2KOazxSdDD8U3MDVm4oLeQhGmD" />
     <add key="okta:OktaDomain" value="https://dev-827480.okta.com" />
+    <add key="okta:Token" value="00jBKM1iYzrgego1vighXrxLNhtJPmX2TeOAhQ-tO5" />
 
     <add key="okta:RedirectUri" value="https://localhost:44396/authorization-code/callback" />
     <add key="okta:PostLogoutRedirectUri" value="https://localhost:44396/Account/PostLogout" />
+    
   </appSettings>
   <system.web>
     <compilation debug="true" targetFramework="4.7.2" />
